requirements.txt
# Data stack
numpy
pandas
polars
pyarrow

# Excel I/O
fastexcel
openpyxl
xlsxwriter

# SQL engine
duckdb

# Frontend
pyside6

# Validation and schemas
jsonschema
pydantic
pyyaml

# Packaging and tooling
pathspec
pre-commit
pyinstaller
python-dateutil
setuptools
wheel
